Change Encoder Steps to Float
The encoderSteps and zEncoderSteps variables were changed from int to float, and the changeEncoderResolution function updated accordingly.

+++ cnc_ctrl_v1/Axis.cpp
@@ -184,7 +184,7 @@ float  Axis::getPitch(){
     return _mmPerRotation;
 }
 
-void   Axis::changeEncoderResolution(const int& newResolution){
+void   Axis::changeEncoderResolution(const float& newResolution){
     /*
     Reassign the encoder resolution for the axis.
     */
